fixadd start and stop button for each service
